% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_label.R
\name{append_label}
\alias{append_label}
\title{Append to \verb{@label} of \linkS4class{panel}}
\usage{
append_label(x, by)
}
\arguments{
\item{x}{a \linkS4class{panel}}

\item{by}{\link[base]{character} scalar}
}
\description{
Append to \verb{@label} of \linkS4class{panel}
}
\note{
There is no generic function \verb{labels<-} in package \pkg{base} !!!

Function \link[base]{comment<-} is not an S3 generic.
}
