% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{as_flextable.panellist}
\alias{as_flextable.panellist}
\title{as_flextable.panellist}
\usage{
\method{as_flextable}{panellist}(x, ...)
}
\arguments{
\item{x}{a \link{panellist}}

\item{...}{..}
}
\description{
as_flextable.panellist
}
\keyword{internal}
