% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{autoplot.panellist}
\alias{autoplot.panellist}
\alias{autolayer.panellist}
\title{Visualize \link{panellist}}
\usage{
\method{autoplot}{panellist}(object, ...)

\method{autolayer}{panellist}(object, ...)
}
\arguments{
\item{object}{a \link{panellist}}

\item{...}{additional parameters, currently no use}
}
\description{
Visualize \link{panellist}
}
\keyword{internal}
