% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.panel.R
\name{consort.panel}
\alias{consort.panel}
\title{\code{\link[=consort.panel]{consort.panel()}}}
\usage{
consort.panel(x, ...)
}
\arguments{
\item{x}{a \linkS4class{panel}}

\item{...}{additional parameters of function \link[consort]{consort_plot},
except for \code{data}, \code{orders} and \code{side_box}}
}
\description{
\code{\link[=consort.panel]{consort.panel()}}
}
\keyword{internal}
