% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.panel.R
\name{consort.panellist}
\alias{consort.panellist}
\title{\code{\link[=consort.panellist]{consort.panellist()}}}
\usage{
consort.panellist(x, ...)
}
\arguments{
\item{x}{a \link{panellist}}

\item{...}{additional parameters for function \link[patchwork]{wrap_plots},
\strong{not} for function \code{\link[=consort.panel]{consort.panel()}}}
}
\description{
\code{\link[=consort.panellist]{consort.panellist()}}
}
\keyword{internal}
