% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{md_.panel}
\alias{md_.panel}
\title{R Markdown Lines for \linkS4class{panel}}
\usage{
\method{md_}{panel}(x, xnm, ...)
}
\arguments{
\item{x}{a \linkS4class{panel}}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Lines for \linkS4class{panel}
}
\keyword{internal}
