% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{minPanel-package}
\alias{minPanel}
\alias{minPanel-package}
\title{minPanel: Aejaz}
\description{
To perform advanced manipulation of R packages.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} [copyright holder]

}
\keyword{internal}
