% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{names<-.panel}
\alias{names<-.panel}
\title{Set \verb{@label} for \linkS4class{panel}}
\usage{
\method{names}{panel}(x) <- value
}
\arguments{
\item{x}{a \linkS4class{panel}}

\item{value}{\link[base]{character} scalar}
}
\description{
Set \verb{@label} for \linkS4class{panel}
}
\note{
There is no generic function \verb{labels<-} in package \pkg{base} !!!

Function \link[base]{comment<-} is not an S3 generic.
}
