% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\docType{class}
\name{panel}
\alias{panel}
\alias{panel-class}
\title{\linkS4class{panel}}
\usage{
panel(m1, m0)
}
\arguments{
\item{m1, m0}{see \strong{Slots}}
}
\description{
\linkS4class{panel}
}
\section{Slots}{

\describe{
\item{\code{m1}}{\link[base]{logical} \link[base]{matrix}, true positives,
i.e., variants tested positive in positive subjects (patients)
Rows are different variants.  Columns are different subjects.}

\item{\code{m0}}{\link[base]{logical} \link[base]{matrix}, false positives,
i.e., variants tested positive in negative subjects (patients)
Rows are different variants.  Columns are different subjects.
Order of variants in \code{m0} and \code{m1} must be the same}

\item{\code{id}}{\link[base]{list} of \link[base]{character} \link[base]{vector}s}

\item{\code{label}}{\link[base]{character} scalar}
}}

\keyword{internal}
